(REF) Extract UpgraderRevisionsTrait
